+++ platforms/posix/cmake/sitl_tests.cmake
@@ -15,7 +15,6 @@ set(tests
 	file2
 	float
 	hrt
-	hysteresis
 	int
 	IntrusiveQueue
 	List
@@ -40,7 +39,6 @@ set(tests
 
 if (${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
 	list(REMOVE_ITEM tests
-		hysteresis
 		mixer
 		uorb
 	)
@@ -48,7 +46,6 @@ endif()
 
 if (CMAKE_SYSTEM_NAME STREQUAL "CYGWIN")
 	list(REMOVE_ITEM tests
-		hysteresis # Intermittent timing fails.
 		uorb
 	)
 endif()
