+++ src/lib/CMakeLists.txt
@@ -45,6 +45,7 @@ add_subdirectory(conversion)
 add_subdirectory(drivers)
 add_subdirectory(ecl)
 add_subdirectory(FlightTasks)
+add_subdirectory(hysteresis)
 add_subdirectory(landing_slope)
 add_subdirectory(led)
 add_subdirectory(mathlib)
