+++ src/lib/hysteresis/hysteresis.cpp
