+++ src/lib/hysteresis/hysteresis.h
