+++ src/lib/systemlib/CMakeLists.txt
@@ -35,7 +35,6 @@ set(SRCS
 	conversions.c
 	cpuload.c
 	crc.c
-	hysteresis/hysteresis.cpp
 	mavlink_log.c
 	otp.c
 	)
