+++ src/modules/land_detector/CMakeLists.txt
@@ -43,5 +43,6 @@ px4_add_module(
 		VtolLandDetector.cpp
 		RoverLandDetector.cpp
 	DEPENDS
+		hysteresis
 	)
