+++ src/systemcmds/tests/tests_main.h
@@ -55,7 +55,6 @@ extern int test_file2(int argc, char *argv[]);
 extern int test_float(int argc, char *argv[]);
 extern int test_hott_telemetry(int argc, char *argv[]);
 extern int test_hrt(int argc, char *argv[]);
-extern int test_hysteresis(int argc, char *argv[]);
 extern int test_int(int argc, char *argv[]);
 extern int test_IntrusiveQueue(int argc, char *argv[]);
 extern int test_jig_voltages(int argc, char *argv[]);
